feat: Refactor all endpoint except item
- Move GraphQL stuff to related modules
- Use modules instead of packages for endpoints
- Move core to __init__
- Improve code quality in general
