+++ src/ikea_api/endpoints/__init__.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+from ikea_api.endpoints.auth import get_authorized_token, get_guest_token
 from ikea_api.endpoints.cart import Cart
 from ikea_api.endpoints.item.item_ingka import fetch as _fetch_items_specs_ingka
 from ikea_api.endpoints.item.item_iows import fetch as _fetch_items_specs_iows
@@ -8,7 +9,7 @@ from ikea_api.endpoints.order_capture import OrderCapture
 from ikea_api.endpoints.purchases import Purchases
 
 
-class fetch_items_specs:
+class Items:
     @staticmethod
     def ingka(items: str | list[str]):
         return _fetch_items_specs_ingka(items)
@@ -22,4 +23,11 @@ class fetch_items_specs:
         return _fetch_items_specs_pip(items)
 
 
-__all__ = ["Cart", "OrderCapture", "Purchases", "fetch_items_specs"]
+__all__ = [
+    "Cart",
+    "OrderCapture",
+    "Purchases",
+    "Items",
+    "get_guest_token",
+    "get_authorized_token",
+]
