--- src/ikea_api/endpoints/cart/queries.py
@@ -1,14 +1,0 @@
-from ikea_api.endpoints.cart.fragments import cart_props
-
-cart = """
-  query Cart(
-    $languageCode: String
-    ) {
-    cart(languageCode: $languageCode) {
-      ...CartProps
-    }
-  }
-  %s
-""" % (
-    cart_props
-)
